[BTAT-132] Refactored the Test Specifications, moved Session Timeout Controller to its own package
