+++ conf/app.routes
@@ -9,4 +9,4 @@ POST        /feedback                   controllers.FeedbackController.submit
 GET         /thankyou                   controllers.FeedbackController.thankyou
 
 #Timeout Routes
-GET         /session-timeout            controllers.SessionTimeoutController.timeout
+GET         /session-timeout            controllers.timeout.SessionTimeoutController.timeout
